fix(calib): bracket ceiling selector + offline re-evaluation (v2.1)
The v2 MEASUREMENT is sound; the verdict logic was not. Field evidence
(dow_dynamo_disagg_p12, fresh stack, healthy engine): rate 0.25 failed ONLY
latency_climb by +0.04 over ~130 requests (low-rate estimator noise on
long-tailed latencies), 0.5 passed everything, 1.0 was the genuine knee
(+1.43 climb, p99 54s). A clean stable point existed with the knee bracketed
above it, yet the old contiguous-stable-prefix selector -- which required
stability from the LOWEST rate upward -- returned no_stable_point, and its
suggested grid pointed DOWN (lower rates make climb noise worse, not better).

1. BRACKET SELECTOR (select_ceiling): the ceiling is the HIGHEST rate that
   passes every criterion; a failing rate above it brackets the knee -> ok;
   highest-passing is the top of grid -> did_not_saturate; nothing passes ->
   no_stable_point. A sub-threshold failure BELOW the ceiling is recorded and
   logged as a low_rate_anomaly but does not disqualify. No individual
   criterion is weakened. Rows are annotated in place (stable / failed_criteria
   / climb_inconclusive / low_rate_anomaly).
2. CLIMB ESTIMATOR robustness: the climb criterion is gated on a minimum
   completed-sample count (--climb-min-samples, default 30; inconclusive-pass
   + flagged below it). Chosen as a SELECTOR gate on the recorded n_ok, NOT a
   mean->median measurement change: median would change the recorded metric
   (breaking the v2.1 selector-only guarantee) and is not recomputable from the
   stored rows, whereas the gate uses only recorded scalars. p12's genuine
   +1.43 climb (hundreds of samples) is still caught.
3. reeval_calibration.py: re-verdicts recorded calibration JSONs from their own
   sweep rows with the fixed selector, recomputes ceiling/rate_calibrated
   (fraction x achieved-at-ceiling), stamps calibration_method_version=2.1 +
   selector_version + a reevaluation block, republishes atomically, prints
   old->new. Skips engine_failure (needs a fresh sweep), v1 (biased
   measurement), and no-rows files; preserves measurements + provenance.
   --dry-run writes nothing. Re-verdicts the running pass without GPU-hours.
4. Versioning: calibration_method_version is now 2.1 (integer part =
   measurement v2 unchanged, fractional = bracket selector).
   check_calibration_provenance is float-aware -- the hard min-gate accepts 2
   and 2.1; the soft warn compares only the integer/measurement part, so a v2.1
   file does not spuriously warn (dow_campaign.yaml min=2 needs no change).
   suggest_next_grid points DOWN only when the lowest rate fails on a saturation
   signal (drops/ratio/p99), never on climb noise.

Tests: p12's exact recorded sweep -> ok at the 0.5 ceiling; all-fail ->
no_stable_point; top-of-grid -> did_not_saturate; low-sample climb gate
(inconclusive-pass) vs sufficient-sample (still catches climb); v2.1 accepted
by the min-2 gate and silent under the warn; climb-noise gives no downward
grid; reeval republish/dry-run/skip paths. Full suite: 448 pass.

+++ scripts/launch_cell.py
@@ -900,11 +900,13 @@ class CalibrationError(Exception):
 # month-3 run is exactly the staleness this gate exists to stop.
 DEFAULT_CALIBRATION_MAX_AGE_DAYS = 14
 
-# The calibration measurement method this build EXPECTS (mirrors
-# calibrate_rate.CALIBRATION_METHOD_VERSION). A file recorded by an older method
-# is not rejected -- host/image/age still gate it -- but callers can ask
-# check_calibration_provenance to WARN so a v1 (finite-window-biased) ceiling is
-# not silently reused after the v2 fix. Bump in lockstep with calibrate_rate.
+# The calibration MEASUREMENT method this build EXPECTS -- the INTEGER part of
+# calibrate_rate.CALIBRATION_METHOD_VERSION (a selector-only revision like 2 ->
+# 2.1 keeps the same measurement, so it must not trip the soft warn; see the
+# integer-part comparison in check_calibration_provenance). A file recorded by an
+# older method is not rejected -- host/image/age still gate it -- but callers can
+# ask check_calibration_provenance to WARN so a v1 (finite-window-biased) ceiling
+# is not silently reused after the v2 fix. Bump when the MEASUREMENT changes.
 EXPECTED_CALIBRATION_METHOD_VERSION = 2
 
 
@@ -974,20 +976,27 @@ def check_calibration_provenance(calib: dict, current_sig: dict,
         v1 finite-window ceilings this campaign is re-taking as v2).
       - warn_method_version (SOFT): a file whose version differs is only warned
         about on stderr. Ignored when min_method_version already covers it."""
+    # Parse as a FLOAT so a selector-only revision (2 -> 2.1: same measurement,
+    # different verdict logic) is representable and ordered correctly. The integer
+    # part is the MEASUREMENT version; the fractional part is the selector revision
+    # (see calibrate_rate.CALIBRATION_METHOD_VERSION).
     rec_ver_raw = calib.get("calibration_method_version", 1)
     try:
-        rec_ver = int(rec_ver_raw)
+        rec_ver = float(rec_ver_raw)
     except (TypeError, ValueError):
         rec_ver = None  # unparseable -> treat as unknown/too-old below
     if min_method_version is not None:
-        if rec_ver is None or rec_ver < int(min_method_version):
+        if rec_ver is None or rec_ver < float(min_method_version):
             raise CalibrationError(
                 f"calibration_method_version={rec_ver_raw!r} is below the required "
                 f"minimum v{min_method_version}: this ceiling was measured with a "
                 "superseded method (e.g. the v1 finite-window sweep). Re-run "
                 "scripts/calibrate_rate.py to re-calibrate with the current method.")
     elif warn_method_version is not None:
-        if rec_ver is None or rec_ver != int(warn_method_version):
+        # Warn only on a MEASUREMENT-version mismatch (integer part): a v2.1 file
+        # has the same measurement as the expected v2, so it must NOT warn -- only
+        # the biased v1 (or an unparseable version) should.
+        if rec_ver is None or int(rec_ver) != int(warn_method_version):
             print(
                 f"[launch_cell] WARNING: calibration_method_version={rec_ver_raw!r} "
                 f"differs from the expected v{warn_method_version}; this ceiling "
